Defer requests to OAuth Session
Previously, most requests were not being handled by the Oauth1
Session that was being created for authenticated users, preventing
us from calling API methods that require authentication. Now, all
methods go through this authenticated session if it exists, and
through the old request path if not. This means that if the user
does not authenticate itself, the functionality will work as
it did before this commit.